streamer/pyproject.toml
[project]
name = "gachibass"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "pygobject>=3.52.3",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.3",
    "pyyaml>=6.0.2",
    "uvicorn>=0.35.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
